+++ src/Ki/Internal.hs
@@ -4,9 +4,7 @@ module Ki.Internal
     module Ki.Ctx,
     module Ki.Duration,
     module Ki.Scope,
-    module Ki.ScopeClosing,
     module Ki.Thread,
-    module Ki.ThreadFailed,
     module Ki.Timeout,
   )
 where
@@ -16,7 +14,5 @@ import Ki.Context
 import Ki.Ctx
 import Ki.Duration
 import Ki.Scope
-import Ki.ScopeClosing
 import Ki.Thread
-import Ki.ThreadFailed
 import Ki.Timeout