+++ src/Ki/Thread.hs
@@ -1,5 +1,5 @@
 module Ki.Thread
-  ( Thread,
+  ( Thread (..),
     async,
     asyncWithUnmask,
     await,
@@ -13,21 +13,16 @@ module Ki.Thread
 where
 
 import Control.Exception (Exception (fromException))
-import Data.Bifunctor (first)
 import qualified Ki.Context as Context
 import Ki.Duration (Duration)
 import Ki.Prelude
 import Ki.Scope (Scope (Scope))
 import qualified Ki.Scope as Scope
-import Ki.ScopeClosing (ScopeClosing (ScopeClosing))
-import Ki.ThreadFailed (ThreadFailed (ThreadFailed), ThreadFailedAsync (ThreadFailedAsync))
 import Ki.Timeout (timeoutSTM)
 
 -- | A running __thread__.
-data Thread a = Thread
-  { threadId :: !ThreadId,
-    action :: !(STM a)
-  }
+data Thread a
+  = Thread !ThreadId !(STM a)
   deriving stock (Functor, Generic)
 
 instance Eq (Thread a) where
@@ -38,12 +33,12 @@ instance Ord (Thread a) where
   compare (Thread id1 _) (Thread id2 _) =
     compare id1 id2
 
--- | Create a __thread__ within a __scope__ to compute a value concurrently.
+-- | Create a __thread__ within a __scope__.
 --
 -- /Throws/:
 --
 --   * Calls 'error' if the __scope__ is /closed/.
-async :: Scope -> IO a -> IO (Thread (Either ThreadFailed a))
+async :: Scope -> IO a -> IO (Thread (Either SomeException a))
 async scope action =
   asyncWithRestore scope \restore -> restore action
 
@@ -52,41 +47,35 @@ async scope action =
 -- /Throws/:
 --
 --   * Calls 'error' if the __scope__ is /closed/.
-asyncWithUnmask :: Scope -> ((forall x. IO x -> IO x) -> IO a) -> IO (Thread (Either ThreadFailed a))
+asyncWithUnmask :: Scope -> ((forall x. IO x -> IO x) -> IO a) -> IO (Thread (Either SomeException a))
 asyncWithUnmask scope action =
   asyncWithRestore scope \restore -> restore (action unsafeUnmask)
 
-asyncWithRestore :: forall a. Scope -> ((forall x. IO x -> IO x) -> IO a) -> IO (Thread (Either ThreadFailed a))
+asyncWithRestore :: forall a. Scope -> ((forall x. IO x -> IO x) -> IO a) -> IO (Thread (Either SomeException a))
 asyncWithRestore scope action = do
   resultVar <- newEmptyTMVarIO
-  childThreadId <-
-    Scope.scopeFork scope action \childThreadId result ->
-      putTMVarIO resultVar (first (ThreadFailed childThreadId) result)
+  childThreadId <- Scope.scopeFork scope action (putTMVarIO resultVar)
   pure (Thread childThreadId (readTMVar resultVar))
 
+-- | Wait for a __thread__ to finish.
 await :: Thread a -> IO a
 await =
   atomically . awaitSTM
 
 -- | @STM@ variant of 'await'.
 awaitSTM :: Thread a -> STM a
-awaitSTM Thread {action} =
+awaitSTM (Thread _ action) =
   action
 
 -- | Variant of 'await' that gives up after the given duration.
---
--- @
--- 'awaitFor' thread duration =
---   'timeout' duration (pure . Just \<$\> 'awaitSTM' thread) (pure Nothing)
--- @
 awaitFor :: Thread a -> Duration -> IO (Maybe a)
 awaitFor thread duration =
   timeoutSTM duration (pure . Just <$> awaitSTM thread) (pure Nothing)
 
--- | Create a __thread__ within a __scope__ to compute a value concurrently.
+-- | Create a __thread__ within a __scope__.
 --
--- If the __thread__ throws an exception, the exception is wrapped in 'ThreadFailed' and immediately propagated up the
--- call tree to the __thread__ that opened its __scope__.
+-- If the __thread__ throws an exception, the exception is immediately propagated up the call tree to the __thread__
+-- that opened its __scope__.
 --
 -- /Throws/:
 --
@@ -127,40 +116,45 @@ forkWithRestore scope action = do
   parentThreadId <- myThreadId
   resultVar <- newEmptyTMVarIO
   childThreadId <-
-    Scope.scopeFork scope action \childThreadId -> \case
-      Left exception -> do
-        whenM
-          (shouldPropagateException scope exception)
-          (throwTo parentThreadId (ThreadFailedAsync threadFailedException))
-        putTMVarIO resultVar (Left threadFailedException)
-        where
-          threadFailedException :: ThreadFailed
-          threadFailedException =
-            ThreadFailed childThreadId exception
-      Right result -> putTMVarIO resultVar (Right result)
-  pure (Thread childThreadId (readTMVar resultVar >>= either throwSTM pure))
+    Scope.scopeFork scope action \case
+      Left exception ->
+        -- Intentionally don't fill the result var.
+        --
+        -- Prior to 0.2.0, we did put a 'Left exception' in the result var, so that if another thread awaited it, we'd
+        -- promptly deliver them the exception that brought this thread down. However, that exception was *wrapped* in
+        -- a 'ThreadFailed' exception, so the caller could distinguish between async exceptions *delivered to them* and
+        -- async exceptions coming *synchronously* out of the call to 'await'.
+        --
+        -- At some point I reasoned that if one is following some basic structured concurrency guidelines, and not doing
+        -- weird/complicated things like passing threads around, then it is likely that a failed forked thread is just
+        -- about to propagate its exception to all callers of 'await' (presumably, its direct parent).
+        --
+        -- Might GHC deliver a BlockedIndefinitelyOnSTM in the meantime, though?
+        maybePropagateException scope parentThreadId exception
+      Right result -> putTMVarIO resultVar result
+  pure (Thread childThreadId (readTMVar resultVar))
 
 forkWithRestore_ :: Scope -> ((forall x. IO x -> IO x) -> IO ()) -> IO ()
 forkWithRestore_ scope action = do
   parentThreadId <- myThreadId
-  _childThreadId <-
-    Scope.scopeFork scope action \childThreadId ->
-      onLeft \exception -> do
-        whenM
-          (shouldPropagateException scope exception)
-          (throwTo parentThreadId (ThreadFailedAsync (ThreadFailed childThreadId exception)))
+  _childThreadId <- Scope.scopeFork scope action (onLeft (maybePropagateException scope parentThreadId))
   pure ()
 
-shouldPropagateException :: Scope -> SomeException -> IO Bool
-shouldPropagateException Scope {closedVar, context} exception =
-  case fromException exception of
-    -- Our scope is (presumably) closing, so don't propagate this exception that presumably just came from our parent.
-    -- But if our scope's closedVar isn't True, that means this 'ScopeClosing' definitely came from somewhere else...
-    Just ScopeClosing -> not <$> readTVarIO closedVar
-    Nothing ->
+maybePropagateException :: Scope -> ThreadId -> SomeException -> IO ()
+maybePropagateException Scope {closedVar, context} parentThreadId exception =
+  whenM shouldPropagateException (throwTo parentThreadId (Scope.ThreadFailed exception))
+  where
+    shouldPropagateException :: IO Bool
+    shouldPropagateException =
       case fromException exception of
-        -- We (presumably) are honoring our own cancellation request, so don't propagate that either.
-        -- It's a bit complicated looking because we *do* want to throw this token if we (somehow) threw it
-        -- "inappropriately" in the sense that it wasn't ours to throw - it was smuggled from elsewhere.
-        Just token -> atomically ((/= token) <$> Context.contextCancelTokenSTM context <|> pure True)
-        Nothing -> pure True
+        -- Our scope is (presumably) closing, so don't propagate this exception that presumably just came from our
+        -- parent. But if our scope's closedVar isn't True, that means this 'ScopeClosing' definitely came from
+        -- somewhere else...
+        Just Scope.ScopeClosing -> not <$> readTVarIO closedVar
+        Nothing ->
+          case fromException exception of
+            -- We (presumably) are honoring our own cancellation request, so don't propagate that either.
+            -- It's a bit complicated looking because we *do* want to throw this token if we (somehow) threw it
+            -- "inappropriately" in the sense that it wasn't ours to throw - it was smuggled from elsewhere.
+            Just token -> atomically ((/= token) <$> Context.contextCancelTokenSTM context <|> pure True)
+            Nothing -> pure True